requirements.txt
# Core ML and Deep Learning
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scipy==1.11.1

# Image Processing
opencv-python==4.8.0.76
Pillow==10.0.0

# OCR Libraries
pytesseract==0.3.10
easyocr==1.7.0

# Translation
googletrans==4.0.0rc1
langdetect==1.0.9

# Data Processing
pandas==2.0.3
matplotlib==3.7.2

# Video Processing
moviepy==1.0.3

# Utilities
tqdm==4.66.1
